ATOMIC_VAR_INIT is deprecated in C17
As per https://en.cppreference.com/w/c/atomic/ATOMIC_VAR_INIT: "This macro was a part of early draft design for C11 atomic types. It is not needed in C11, and is deprecated in C17 and removed in C23". 

This fixes an error that I get when I try to build with "-Wall -Werror" using zig cc.